+++ example/lib/webgl_shadowmap_viewer.dart
@@ -195,7 +195,7 @@ class _MyAppState extends State<webgl_shadowmap_viewer> {
         "format": THREE.RGBAFormat
       });
       renderTarget = THREE.WebGLRenderTarget(
-          (width * dpr), (height * dpr), pars);
+          (width * dpr).toInt(), (height * dpr).toInt(), pars);
       renderTarget.samples = 4;   
       renderer!.setRenderTarget(renderTarget);
       sourceTexture = renderer!.getRenderTargetGLTexture(renderTarget);
